install tweaks and single include file

+++ include/scratchy/BitMask.hpp
@@ -3,6 +3,6 @@
 template <class T>
 struct BitMask
 {
-    static const size_t type_size = sizeof(T) * 8;
+    static const unsigned int type_size = sizeof(T) * 8;
     static unsigned long const value = (static_cast<unsigned long>(1) << type_size) - 1;
 };
